Migrate config to MidnightConfig
This shrinks the jar from 1.2MiB to 360KiB. It accomplishes this by migrating from Cloth Config (which is quite overkill for this mod) to MidnightConfig, by Motschen. This also means the config code will be significantly cleaner.
The main downside to this is that existing configs won't migrate over to the new one (as it uses json instead of toml). I believe this sacrifice is fine, though, since json is easier for data interchange.
I have tested this, and everything works as intended as far as I can tell (even translations).