Shrink icon to 128x128 from 1024x1024
This shrinks the icon image from 302KB to 9KB, according to [PicResize](https://picresize.com). On top of #29, this shrinks the jar down to 78.9KiB.
The reason I've done this is because, realistically, nobody is actually going to see it in the full 1024x1024 quality; 128x128 should be more than enough.