Fix being unable to persistently set certain volume multipliers
Volume multipliers which had a default value !- 100% were not saved
to disk, and would thus revert to their default on subsequent startups.

As of this commit each config now knows which power state it's for,
and can compare whether it should save a volume override of 100% propery.

Fixes https://github.com/juliand665/Dynamic-FPS/issues/199.